update RN to 0.57.3, added iOS config, build scripts
