+++ package.json
@@ -6,22 +6,33 @@
 		"start": "node node_modules/react-native/local-cli/cli.js start",
 		"test": "jest",
 		"appium": "appium",
-		"appium:doctor": "appium-doctor"
+		"android": "react-native run-android",
+		"ios": "react-native run-ios",
+		"appium:doctor": "appium-doctor",
+		"build-android-debug": "cd android; ./gradlew assembleDebug",
+		"build-android-release": "cd android; ./gradlew assembleRelease",
+		"build-ios-debug": "cd ios; xcodebuild -UseModernBuildSystem=NO -scheme 'appiumTutorial' -sdk iphonesimulator -configuration Debug -derivedDataPath build/app; zip -r app-debug.zip build/app/Build/Products/Debug-iphonesimulator/*.app",
+		"build-ios-release": "cd ios; xcodebuild -UseModernBuildSystem=NO -scheme 'appiumTutorial' -sdk iphonesimulator -configuration Release -derivedDataPath build/app; zip -r app-release.zip build/app/Build/Products/Release-iphonesimulator/*.app"
 	},
 	"dependencies": {
-		"react": "16.0.0-alpha.12",
-		"react-native": "0.48.2"
+		"react": "^16.6.0-alpha.8af6728",
+		"react-native": "^0.57.3"
 	},
 	"devDependencies": {
-		"appium": "^1.9.0",
-		"appium-doctor": "^1.4.3",
-		"babel-jest": "21.0.2",
-		"babel-preset-react-native": "3.0.2",
+		"@babel/core": "^7.0.0",
+		"@babel/runtime": "^7.0.0",
+		"@babel/preset-env": "^7.1.0",
+		"appium": "^1.9.2-beta.2",
+		"appium-doctor": "^1.6.0",
+		"babel-jest": "23.6.0",
+		"babel-preset-react-native": "4.0.1",
 		"jest": "^23.6.0",
-		"react-test-renderer": "16.0.0-alpha.12",
-		"wd": "^1.4.0"
+		"metro-react-native-babel-preset": "0.48.1",
+		"react-test-renderer": "^16.6.0-alpha.8af6728",
+		"wd": "^1.11.0"
 	},
 	"jest": {
-		"preset": "react-native"
+		"preset": "react-native",
+		"transform": { "^.+\\.js$": "<rootDir>/node_modules/react-native/jest/preprocessor.js" }
 	}
 }
